# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PYX Gate API
  version: "1.0.0"
  description: |
    API pública da PYX Gate para criação de cobranças Pix e cartão,
    consulta de pagamentos e recebimento de eventos via webhook.

    Autenticação por chave de API (`Authorization: Bearer sk_...`). Toda
    transação criada com uma chave `sk_test_...` roda em **modo teste**:
    nunca move dinheiro real, nunca chama a adquirente de verdade — a
    confirmação é feita manualmente via `/v1/test/payments/{id}/pay` (ou
    `/fail`). Chaves `sk_live_...` processam pagamentos reais.
  contact:
    name: PYX Gate
servers:
  - url: http://localhost:3000/v1
    description: Ambiente local de desenvolvimento
  - url: https://api.pyxgate.com/v1
    description: Produção (domínio ilustrativo — ainda não implantado)
security:
  - bearerAuth: []
...
paths:
...
  /webhook_endpoints:
    post:
      operationId: createWebhookEndpoint
      tags: [Webhook Endpoints]
      summary: Cadastra um endpoint para receber eventos.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookEndpointRequest"
            examples:
              default:
                value:
                  url: "https://seusite.com/webhooks/pyxgate"
                  events: ["payment.paid", "payment.failed"]
      responses:
        "201":
          description: Endpoint criado. O campo `secret` (whsec_...) só é útil aqui e na listagem — guarde para verificar assinaturas.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "400":
          $ref: "#/components/responses/InvalidRequest"
    get:
      operationId: listWebhookEndpoints
      tags: [Webhook Endpoints]
      summary: Lista os endpoints cadastrados.
      responses:
        "200":
          description: Lista de endpoints.
          content:
            application/json:
              schema:
                type: object
                properties:
                  object: { type: string, example: list }
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/WebhookEndpoint"
  /webhook_endpoints/{id}:
    patch:
      operationId: updateWebhookEndpoint
      tags: [Webhook Endpoints]
      summary: Atualiza url, eventos ou ativa/desativa um endpoint.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, example: "we_654a1..." }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string, format: uri }
                events:
                  type: array
                  items: { type: string }
                active: { type: boolean }
      responses:
        "200":
          description: Endpoint atualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteWebhookEndpoint
      tags: [Webhook Endpoints]
      summary: Remove um endpoint.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removido.
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "sk_live_... / sk_test_..."
      description: |
        Secret key gerada no dashboard (Desenvolvedores → Chaves de API).
        Publishable keys (`pk_...`) não autenticam nenhum endpoint desta API
        — são só para uso client-side em SDKs futuros.
...
  responses:
    InvalidRequest:
      description: Payload inválido ou requisição malformada.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Recurso não encontrado (ou não pertence à conta autenticada).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            type:
              type: string
              enum: [invalid_request_error, authentication_error, card_error, api_error]
            code:
              type: string
              example: invalid_api_key
            message:
              type: string
            param:
              type: string
              nullable: true
...
    CreateWebhookEndpointRequest:
      type: object
      required: [url, events]
      properties:
        url:
          type: string
          format: uri
          description: Deve usar https:// em produção.
        events:
          type: array
          items:
            type: string
            enum: ["*", payment.created, payment.paid, payment.failed, payment.expired, refund.succeeded]
    WebhookEndpoint:
      type: object
      properties:
        id: { type: string, example: "we_654a1f2b3c4d5e6f70718293" }
        url: { type: string }
        events:
          type: array
          items: { type: string }
        secret:
          type: string
          description: "whsec_... — use para verificar a assinatura PYX-Signature."
        active: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }